blueprint-manager/src/test/java/com/sequenceiq/cloudbreak/blueprint/BlueprintSegmentReaderTest.java
package com.sequenceiq.cloudbreak.blueprint;

import java.io.IOException;
import java.util.Map;

import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.runners.MockitoJUnitRunner;
import org.springframework.test.util.ReflectionTestUtils;

import com.sequenceiq.cloudbreak.blueprint.templates.ServiceName;
import com.sequenceiq.cloudbreak.blueprint.templates.TemplateFiles;

@RunWith(MockitoJUnitRunner.class)
public class BlueprintSegmentReaderTest {

    @InjectMocks
    private final BlueprintSegmentReader underTest = new BlueprintSegmentReader();

    @Before
    public void setup() throws IOException {
        ReflectionTestUtils.setField(underTest, "blueprintTemplatePath", "blueprints");
        ReflectionTestUtils.setField(underTest, "basicTemplatePath", "basics");
    }

    @Test
    public void testThatAllFileIsReadableShouldVerifyThatFileCountMatch() {
        Map<ServiceName, TemplateFiles> configFiles = underTest.collectAllConfigFile();
        Map<ServiceName, TemplateFiles> serviceFiles = underTest.collectAllServiceFile();

        Assert.assertEquals(configFiles.size(), 1);
        Assert.assertEquals(serviceFiles.size(), 9);
    }
}